backend/auth/requirements.txt
altgraph @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot11/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-141/altgraph-0.17.2-py2.py3-none-any.whl
appnope==0.1.3
asgiref==3.7.2
astroid==3.0.3
asttokens==2.4.0
attrs==23.1.0
auth==0.5.3
backcall==0.2.0
bcrypt==4.1.1
blinker==1.7.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
comm==0.1.4
confusable-homoglyphs==3.2.0
contourpy==1.2.0
cryptography==41.0.5
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
dj-rest-auth==5.0.2
Django==4.2.6
django-cors-headers==4.3.1
django-countries==7.5.1
django-filter==23.4
django-phone-auth==0.3.1
django-phone-field==1.8.1
django-phonenumber-field==7.2.0
django-registration==3.4
django-rest-auth==0.9.5
django-rest-auth-forked==0.9.5
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dnspython==2.4.2
docutils==0.20.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
eventlet==0.34.2
exceptiongroup==1.1.3
executing==1.2.0
falcon==3.1.3
fonttools==4.45.0
future @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot11/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-141/future-0.18.2-py3-none-any.whl
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
inflection==0.5.1
ipykernel==6.25.2
ipython==8.15.0
isort==5.13.2
jedi==0.19.0
joblib==1.3.2
jupyter_client==8.3.1
jupyter_core==5.3.1
kiwisolver==1.4.5
macholib @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot11/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-141/macholib-1.15.2-py2.py3-none-any.whl
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mongoengine==0.27.0
mypy==1.8.0
mypy-extensions==1.0.0
mysql-connector-python==8.0.33
nest-asyncio==1.5.8
numpy==1.26.2
oauthlib==3.2.2
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==23.1
pandas==2.1.3
parso==0.8.3
pexpect==4.8.0
phonenumbers==8.13.27
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
pylint==3.0.3
pymongo==4.6.1
pyotp==2.9.0
pyparsing==3.1.1
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
requests==2.31.0
requests-oauthlib==1.3.1
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
selenium==4.15.2
Send2Trash==1.8.2
six @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot11/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-141/six-1.15.0-py2.py3-none-any.whl
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
stack-data==0.6.2
thonny==4.1.4
threadpoolctl==3.2.0
tomli==2.0.1
tomlkit==0.12.3
tornado==6.3.3
traitlets==5.10.0
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
wcwidth==0.2.6
webdriver-manager==4.0.1
Werkzeug==3.0.1
wsproto==1.2.0
zipp==3.17.0
